chore(home): Fix css for TopAndSidebar component
