--- components/common/Top-and-SideBar/TopAndSideBar.html
@@ -1,112 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" >
-<head>
-  <meta charset="UTF-8">
-  <title>CodePen - Improved Bar</title>
-  <link rel='stylesheet' href='https://cdnjs.cloudflare.com/ajax/libs/tailwindcss/2.2.19/tailwind.min.css'>
-
-</head>
-<body>
-<!-- partial:index.partial.html -->
-<script defer src="https://unpkg.com/alpinejs@3.2.4/dist/cdn.min.js"></script>
-<div class="antialiased min-h-screen bg-blue-900" x-data="{open: true}">
-  <nav class="absolute transform duration-200 z-10 w-96 bg-white text-white min-h-screen p-3"
-       :class="{'translate-x-0 ease-in':open === true, '-translate-x-full ease-out opacity-0': open === false}">
-    <div class="flex justify-between">
-        <span class="font-bold text-2xl p-2">SideBar</span>
-        <button class="p-2 text-blue-900 focus:outline-none focus:bg-blue-400 hover:bg-blue-400 rounded-md"
-                :class="{'opacity-0':open === false, 'opacity-100': open === true}"
-                @click="open = false">
-          <svg xmlns="http://www.w3.org/2000/svg" class="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor" stroke-width="2">
-  <path stroke-linecap="round" stroke-linejoin="round" d="M4 6h16M4 12h16M4 18h16" />
-</svg>
-        </button>
-    </div>
-    <!-- Circle -->
-    <div class="pt-72">
-        <section x-data="DegreeLeft"
-        class="">
-
-        <div class="flex items-center justify-center pr-24" x-data="{ circumference: 2 * 22 / 7 * 60 }">
-            <svg class="transform -rotate-90 w-72 h-64">
-                <circle cx="145" cy="145" r="60" stroke="currentColor" stroke-width="15" fill="transparent"
-                    class="text-blue-900" />
-
-                <circle cx="145" cy="145" r="60" stroke="currentColor" stroke-width="15" fill="transparent"
-                    :stroke-dasharray="circumference"
-                    :stroke-dashoffset="circumference - currentSkill.percent / 100 * circumference"
-                    stroke-linecap="round"
-                    class="text-green-300" />
-            </svg>
-          <!--No Text Shadow, will have to add -->
-            <span class="absolute text-4xl text-blue-900 font-semibold font-sans pb-8 pl-8 drop-shadow-xl" x-text="`${currentSkill.percent}%`"></span>
-          </div>
-    </section>
-    </div>
-    <span class="font-sans text-4xl text-blue-900 font-semibold pl-20">Student Name</span>
-    
-    <!-- Info Bubbles -->
-    <div class=" mt-8 flex flex-col space-y-3">
-    <span class="bg-blue-300 ml-20 rounded-xl font-bold text-black text-center py-1 mx-40 shadow-md">Sophomore</span>
-      
-    <span class="bg-red-200 ml-20 rounded-xl font-bold text-black text-center px-4 py-1 m-auto shadow-md">Cognitive Science B.S.</span>
-      
-    <span class="bg-green-200 ml-20 rounded-xl font-bold text-black text-center px-4 py-1 m-auto shadow-md">Math(minor)</span>
-    </div>
-  </nav>
-  <div class="relative transform duration-200 flex items-center z-0 bg-gray-500 last:align-right"
-       :class="{'translate-x-96 ease-in':open === true, '-translate-x-0 ease-out': open === false}">
-    <button class="p-2 text-white transition-colors hover:bg-gray-600 hover:bg-gray-700"
-            :class="{'opacity-0':open === true, 'opacity-100': open === false}"
-            @click="open = true">
-          <svg xmlns="http://www.w3.org/2000/svg" class="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor" stroke-width="2">
-  <path stroke-linecap="round" stroke-linejoin="round" d="M4 6h16M4 12h16M4 18h16" />
-</svg>
-    </button>
-    <div>
-      <a href="#" class="flex items-center justify-center w-40 px-4 py-4 mt-2 space-x-1 text-md text-white transition-colors hover:bg-gray-600 hover:bg-gray-700 hover:underline">
-        <svg xmlns="http://www.w3.org/2000/svg" class="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor" stroke-width="2">
-  <path stroke-linecap="round" stroke-linejoin="round" d="M3 12l2-2m0 0l7-7 7 7M5 10v10a1 1 0 001 1h3m10-11l2 2m-2-2v10a1 1 0 01-1 1h-3m-6 0a1 1 0 001-1v-4a1 1 0 011-1h2a1 1 0 011 1v4a1 1 0 001 1m-6 0h6" />
-</svg>
-        <span class="text-md font-semibold self-end text-white dark:text-gray-200">Home</span></a>
-    </div>
-    <div>
-      <a href="#" class="flex items-center justify-center w-40 px-4 py-4 mt-2 space-x-1 text-md text-white transition-colors hover:bg-gray-600 hover:bg-gray-700 hover:underline">
-        <svg xmlns="http://www.w3.org/2000/svg" class="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor" stroke-width="2">
-  <path stroke-linecap="round" stroke-linejoin="round" d="M9 12h6m-6 4h6m2 5H7a2 2 0 01-2-2V5a2 2 0 012-2h5.586a1 1 0 01.707.293l5.414 5.414a1 1 0 01.293.707V19a2 2 0 01-2 2z" />
-</svg>
-        <span class="text-md font-semibold self-end text-white dark:text-gray-200">Degrees</span></a>
-    </div>
-    <div>
-      <a href="#" class="flex items-center justify-center w-40 px-4 py-4 mt-2 space-x-1 text-md text-white transition-colors hover:bg-gray-600 hover:bg-gray-700 hover:underline">
-        <svg xmlns="http://www.w3.org/2000/svg" class="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor" stroke-width="2">
-  <path stroke-linecap="round" stroke-linejoin="round" d="M16 7a4 4 0 11-8 0 4 4 0 018 0zM12 14a7 7 0 00-7 7h14a7 7 0 00-7-7z" />
-</svg>
-        <span class="text-md font-semibold self-end text-white dark:text-gray-200">Profile</span></a>
-    </div>
-    <div>
-      <a href="#" class="flex items-center justify-center w-40 px-4 py-4 mt-2 space-x-1 text-md text-white transition-colors hover:bg-gray-600 hover:bg-gray-700 hover:underline">
-        <svg xmlns="http://www.w3.org/2000/svg" class="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor" stroke-width="2">
-  <path d="M12 14l9-5-9-5-9 5 9 5z" />
-  <path d="M12 14l6.16-3.422a12.083 12.083 0 01.665 6.479A11.952 11.952 0 0012 20.055a11.952 11.952 0 00-6.824-2.998 12.078 12.078 0 01.665-6.479L12 14z" />
-<path stroke-linecap="round" stroke-linejoin="round" d="M12 14l9-5-9-5-9 5 9 5zm0 0l6.16-3.422a12.083 12.083 0 01.665 6.479A11.952 11.952 0 0012 20.055a11.952 11.952 0 00-6.824-2.998 12.078 12.078 0 01.665-6.479L12 14zm-4 6v-7.5l4-2.222" />
-</svg>
-        <span class="text-md font-semibold self-end text-white dark:text-gray-200">Credits</span></a> 
-    </div>
-  </div>
-</div>
-
-<script>
-  document.addEventListener('alpine:init', () => {
-    Alpine.data('DegreeLeft', () => ({
-      currentSkill: {
-        'title': 'ClassPercentLeft',
-        'percent': '75',
-      }
-    }));
-  });
-</script>
-<!-- partial -->
-  
-</body>
-</html>
